requirements.txt
alembic==1.18.4
altair==4.2.2
aniso8601==10.0.1
annotated-types==0.7.0
asttokens==3.0.1
attrs==25.4.0
Automat==25.4.16
backcall==0.2.0
backports.zstd==1.3.0
beautifulsoup4==4.14.3
bidict==0.23.1
black==25.12.0
bleach==6.3.0
blinker==1.9.0
boto3==1.42.30
botocore==1.42.30
brotli==1.2.0
cachelib==0.13.0
cachetools==6.2.4
certifi==2025.10.5
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
dash==3.3.0
dask==2026.1.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
distlib==0.4.0
dnspython==2.8.0
docopt==0.6.2
duckdb==1.4.3
email-validator==2.3.0
entrypoints==0.4
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
flake8==7.3.0
Flask==3.1.2
Flask-Caching==2.3.1
Flask-Compress==1.23
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-restx==1.3.2
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
great_expectations==1.11.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hyperlink==21.0.0
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Incremental==24.11.0
iniconfig==2.3.0
ipython==8.12.3
isort==7.0.0
itemadapter==0.13.1
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
limits==5.8.0
locket==1.0.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mccabe==0.7.0
mistune==3.2.0
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.10.0
numpy==2.2.6
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.6
partd==1.4.2
pathspec==1.0.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.1.0
pipreqs==0.5.0
platformdirs==4.5.1
plotly==6.5.2
pluggy==1.6.0
polars==1.37.1
polars-runtime-32==1.37.1
pre_commit==4.5.1
prompt_toolkit==3.0.52
Protego==0.5.0
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyDispatcher==2.0.7
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyparsing==3.3.1
pyspark==4.1.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
queuelib==1.8.0
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
retrying==1.4.2
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.15.3
Scrapy==2.14.1
sentry-sdk==2.54.0
service-identity==24.2.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
streamlit==1.53.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.3.1
toml==0.10.2
tomli==2.4.0
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.5.0
virtualenv==20.36.1
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.5
wrapt==2.1.1
wsproto==1.3.2
WTForms==3.2.1
yarg==0.1.9
zipp==3.23.0
zope.interface==8.2
